+++ src/entity/championship.rs
@@ -2,8 +2,8 @@ use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use sqlx::{FromRow, Type};
 
-#[repr(u8)]
 #[derive(Type, Debug, Serialize, Deserialize)]
+#[sqlx(type_name = "category")]
 pub enum Category {
     F1,
     F2,
@@ -11,13 +11,13 @@ pub enum Category {
 
 #[derive(Debug, Serialize, FromRow)]
 pub struct Championship {
-    pub id: u32,
-    pub port: u16,
+    pub id: i32,
+    pub port: i32,
     pub name: String,
     pub category: Category,
-    pub season: u16,
-    pub driver_count: u8,
-    pub owner_id: u32,
+    pub season: i8,
+    pub driver_count: i16,
+    pub owner_id: i32,
     #[serde(skip_serializing)]
     pub created_at: DateTime<Utc>,
     #[serde(skip_serializing)]
